Added reviewing

+++ bot.py
@@ -19,7 +19,8 @@ bot.remove_command('help')
 
 initial_extensions = ['cogs.applications',
                       'cogs.apply',
-                      'cogs.settings']
+                      'cogs.settings',
+                      'cogs.review']
 
 if __name__ == '__main__':
     for extension in initial_extensions:
@@ -38,52 +39,61 @@ async def on_ready():
 @bot.command(pass_context=True)
 async def reload(ctx, *, msg):
     """Load a module."""
-    await ctx.message.delete()
-    try:
-        if os.path.exists("custom_cogs/{}.py".format(msg)):
-            bot.reload_extension("custom_cogs.{}".format(msg))
-        elif os.path.exists("cogs/{}.py".format(msg)):
-            bot.reload_extension("cogs.{}".format(msg))
+    if ctx.message.author.id ==173450781784145921:
+        await ctx.message.delete()
+        try:
+            if os.path.exists("custom_cogs/{}.py".format(msg)):
+                bot.reload_extension("custom_cogs.{}".format(msg))
+            elif os.path.exists("cogs/{}.py".format(msg)):
+                bot.reload_extension("cogs.{}".format(msg))
+            else:
+                raise ImportError("No module named '{}'".format(msg))
+        except Exception as e:
+            await ctx.send('Failed to reload module: `{}.py`'.format(msg))
+            await ctx.send('{}: {}'.format(type(e).__name__, e))
         else:
-            raise ImportError("No module named '{}'".format(msg))
-    except Exception as e:
-        await ctx.send('Failed to reload module: `{}.py`'.format(msg))
-        await ctx.send('{}: {}'.format(type(e).__name__, e))
+            await ctx.send('Reloaded module: `{}.py`'.format(msg))
     else:
-        await ctx.send('Reloaded module: `{}.py`'.format(msg))
+        await ctx.send('Dev only')
 
 @bot.command(pass_context=True)
 async def unload(ctx, *, msg):
     """Unload a module"""
-    await ctx.message.delete()
-    try:
-        if os.path.exists("cogs/{}.py".format(msg)):
-            bot.unload_extension("cogs.{}".format(msg))
-        elif os.path.exists("custom_cogs/{}.py".format(msg)):
-            bot.unload_extension("custom_cogs.{}".format(msg))
+    if ctx.message.author.id ==173450781784145921:
+        await ctx.message.delete()
+        try:
+            if os.path.exists("cogs/{}.py".format(msg)):
+                bot.unload_extension("cogs.{}".format(msg))
+            elif os.path.exists("custom_cogs/{}.py".format(msg)):
+                bot.unload_extension("custom_cogs.{}".format(msg))
+            else:
+                raise ImportError("No module named '{}'".format(msg))
+        except Exception as e:
+            await ctx.send('Failed to unload module: `{}.py`'.format(msg))
+            await ctx.send('{}: {}'.format(type(e).__name__, e))
         else:
-            raise ImportError("No module named '{}'".format(msg))
-    except Exception as e:
-        await ctx.send('Failed to unload module: `{}.py`'.format(msg))
-        await ctx.send('{}: {}'.format(type(e).__name__, e))
+            await ctx.send('Unloaded module: `{}.py`'.format(msg))
     else:
-        await ctx.send('Unloaded module: `{}.py`'.format(msg))
+        await ctx.send('Dev only')
 
 @bot.command(pass_context=True)
 async def load(ctx, *, msg):
     """Load a module"""
-    await ctx.message.delete()
-    try:
-        if os.path.exists("cogs/{}.py".format(msg)):
-            bot.load_extension("cogs.{}".format(msg))
-        elif os.path.exists("custom_cogs/{}.py".format(msg)):
-            bot.load_extension("custom_cogs.{}".format(msg))
+    if ctx.message.author.id ==173450781784145921:
+        await ctx.message.delete()
+        try:
+            if os.path.exists("cogs/{}.py".format(msg)):
+                bot.load_extension("cogs.{}".format(msg))
+            elif os.path.exists("custom_cogs/{}.py".format(msg)):
+                bot.load_extension("custom_cogs.{}".format(msg))
+            else:
+                raise ImportError("No module named '{}'".format(msg))
+        except Exception as e:
+            await ctx.send('Failed to load module: `{}.py`'.format(msg))
+            await ctx.send('{}: {}'.format(type(e).__name__, e))
         else:
-            raise ImportError("No module named '{}'".format(msg))
-    except Exception as e:
-        await ctx.send('Failed to load module: `{}.py`'.format(msg))
-        await ctx.send('{}: {}'.format(type(e).__name__, e))
+            await ctx.send('Loaded module: `{}.py`'.format(msg))
     else:
-        await ctx.send('Loaded module: `{}.py`'.format(msg))
+        await ctx.send('Dev only')
 
 bot.run('TOKEN')
